move redirect files out of api/auth module, fix imports
